requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
appdirs==1.4.4            # via fs
attrs==19.3.0             # via pytest
certifi==2020.4.5.2       # via requests
chardet==3.0.4            # via requests
coverage==5.1             # via pytest-cov
ddt==1.4.1                # via -r requirements/test.in
django-model-utils==4.0.0  # via -r requirements/base.in
django-waffle==0.20.0     # via edx-django-utils, edx-drf-extensions
drf-jwt==1.14.0           # via edx-drf-extensions
edx-django-utils==3.2.3   # via edx-drf-extensions
edx-drf-extensions==6.0.0  # via -r requirements/base.in
edx-opaque-keys[django]==2.1.0  # via -r requirements/base.in, edx-drf-extensions
factory-boy==2.12.0       # via -r requirements/test.in
faker==4.1.1              # via factory-boy
freezegun==0.3.15         # via -r requirements/test.in
fs==2.4.11                # via xblock
future==0.18.2            # via pyjwkest
idna==2.9                 # via requests
importlib-metadata==1.6.1  # via pluggy, pytest
lxml==4.5.1               # via xblock
markupsafe==1.1.1         # via xblock
mock==3.0.5               # via -c requirements/constraints.txt, -r requirements/test.in
more-itertools==8.4.0     # via pytest
newrelic==5.14.0.142      # via edx-django-utils
packaging==20.4           # via pytest
pathlib2==2.3.5           # via pytest
pbr==5.4.5                # via stevedore
pluggy==0.13.1            # via pytest
psutil==1.2.1             # via edx-django-utils
py==1.8.2                 # via pytest
pycryptodomex==3.9.7      # via pyjwkest
pyjwkest==1.4.2           # via edx-drf-extensions
pyjwt==1.7.1              # via drf-jwt
pymongo==3.10.1           # via edx-opaque-keys
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.0        # via -r requirements/test.in
pytest-django==3.9.0      # via -r requirements/test.in
pytest==5.4.3             # via pytest-cov, pytest-django
python-dateutil==2.8.1    # via edx-drf-extensions, faker, freezegun, xblock
pytz==2020.1              # via -r requirements/base.in, django, fs, xblock
pyyaml==5.3.1             # via xblock
requests==2.23.0          # via edx-drf-extensions, pyjwkest
rest-condition==1.0.3     # via edx-drf-extensions
semantic-version==2.8.5   # via edx-drf-extensions
six==1.15.0               # via django-waffle, edx-drf-extensions, edx-opaque-keys, freezegun, fs, mock, packaging, pathlib2, pyjwkest, python-dateutil, stevedore, xblock
sqlparse==0.3.1           # via django
stevedore==1.32.0         # via edx-opaque-keys
text-unidecode==1.3       # via faker
typing==3.7.4.1           # via fs
urllib3==1.25.9           # via requests
wcwidth==0.2.4            # via pytest
web-fragments==0.3.2      # via xblock
webob==1.8.6              # via xblock
xblock==1.3.1             # via -r requirements/base.in
zipp==1.2.0               # via -c requirements/constraints.txt, importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
